tunnel config: reorganize the trait sealing
Putting the methods in the public trait and leaving Sealed as a no-op
crate-only trait makes for a much better doc viewing experience. The way
it was previously, you couldn't see any of the implementors of
TunnelConfigPrivate, so it was impossible to see what types were
actually valid TunnelConfig implementors.